+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.8]
+
+This update resolves issues with add-ons registering the same Block Entity by implementing a queue system in Platform's Feature Manager. 
+_Add-on developers can register their blocks with Villager Coin and the Feature Manager will register them with their Block Entity at the end of feature initialization._
+
+### Added
+
+- Added `maximumCoinBankCurrencyValue` option to allow controlling the maximum currency value of Coin Banks. 
+_This option is set to the maximum Integer value by default._
+
+### Changed
+
+- Changed supported Platform version to 1.2.4. _This adds queues to the Feature Manager._
+
+### Fixed
+
+- Add-ons now properly load shared block entities as intended.
+
 ## [1.0.7]
 
 This update resolves an issue with modded containers causing a fatal crash when saving stacks greater than 99.